Hash object names before using them as cache filenames
This is fix for CVE-2013-1867, the variation of the fix suggested in
the original CVE report.
    
Note that while the offered patch was essentially correct, the analysis
itself was a little bit mistaken, because it was talking about token
label which would cause vulnerabilities, while in fact the problematic
items are token object (e.g. certificates) friendly names.